topsis-web/requirements.txt
flask
pandas
numpy
yagmail
python-dotenv
